[Modify] #110 OrderDto의 id 타입을 변경

+++ data/src/main/java/com/woowahan/data/entity/dto/OrderDto.kt
@@ -4,6 +4,7 @@ import androidx.room.Embedded
 import androidx.room.Relation
 import com.woowahan.data.entity.table.OrderItemTableEntity
 import com.woowahan.data.entity.table.OrderTableEntity
+import com.woowahan.domain.model.OrderModel
 
 data class OrderDto(
     @Embedded
@@ -22,7 +23,7 @@ data class OrderDto(
 }
 
 data class OrderEntity(
-    val orderId: Int,
+    val orderId: Long,
     val time: String,
     val items: List<OrderItemTableEntity>
 )
